refactor: Appbar 버튼 분리, 폴더 이름 변경
