+++ src/pages/diary/DiaryFlow.tsx
@@ -4,7 +4,7 @@ import { getTodayDate } from "@/utils/util";
 import { useEffect, useState } from "react";
 import { useErrorBoundary } from "react-error-boundary";
 import { useBlocker, useNavigate, useLocation } from "react-router-dom";
-import Appbar from "@/components/common/Appbar";
+import Appbar from "@/components/common/appbar/Appbar";
 import Button from "@/components/common/Button";
 import {
   Dialog,
