+++ src/pages/user/Like.tsx
@@ -1,5 +1,5 @@
 import { getLikedDiaries } from "@/api/api";
-import Appbar from "@/components/common/Appbar";
+import Appbar from "@/components/common/appbar/Appbar";
 import ThumbnailGrid from "@/components/common/ThumbnailGrid";
 import ROUTE_PATH from "@/constants/ROUTE_PATH";
 import useInView from "@/hooks/useInView";
