+++ src/pages/user/Stat.tsx
@@ -1,4 +1,4 @@
-import Appbar from "@/components/common/Appbar";
+import Appbar from "@/components/common/appbar/Appbar";
 import { useState } from "react";
 import { Outlet, useNavigate } from "react-router-dom";
